odm requests debugged and doc-strings provided
