+++ setup.py
@@ -24,6 +24,7 @@ setup(
     python_requires=">=3.8",
     install_requires=[
         "requests",
+        "requests_toolbelt",
         "exifread",
         "Pillow",
         "openpyxl",
